Use ColorScheme map, cleanup Config, add pkg internal/slices
This commit introduces a new ColorScheme map type and a new
ColorSchemeKey enum type that replaces the existing ColorSchemeConfig
structure. This allows for much easier overriding of color schemes.

The new ColorScheme changes also introduces a new colorEnvMap variable
that maps env var prefixes to the respective ColorSchemeKey, which is a
bit cleaner.

The configs are also cleaned up a bit: GuildID is removed, because that
can be obtained from ChannelID, and redundant domain-specific field
names have been cleaned up.

A new package internal/slices is also introduced, which just adds a few
convenient functions for slice-searching and such.